Enable the test suite to run on Windows hosts
There are two changes part of this commit, after which the entire test suite passes correctly.

1) Using the `force` flag when calling `fs.rm`, which is necessary because Windows throws an ENOENT filesystem error (and thus exception) when removing non-existing files and directories. The test suite carefully removes the `generated` directory before each test.
2) Always importing from `path` (rather than `path/poxix`), and correcting the path separators in the two places it matters.

In order to be able to use `String.replaceAll()` I updated the `target` in `tsconfig.json` to es2021, as it seems fairly unlikely that anyone running this script isn't able to run that. An alternative would be to use a regular expression based replace.